+++ lib/td_dd_web/schema/structures.ex
@@ -109,7 +109,8 @@ defmodule TdDdWeb.Schema.Structures do
 
     field :children, list_of(:data_structure_version) do
       arg(:deleted, :boolean, default_value: false)
-      resolve(dataloader(TdDd.DataStructures))
+
+      resolve(&Resolvers.Structures.childrens/3)
     end
 
     field :siblings, list_of(:data_structure_version)