+++ lib/td_dd_web/views/data_structure_view.ex
@@ -173,16 +173,24 @@ defmodule TdDdWeb.DataStructureView do
     Map.put_new(json, :note, latest_note)
   end
 
-  defp maybe_put_my_grant_request(json, %{my_grant_request: [%{} = my_grant_request]}), do:
-    Map.put_new(json, :my_grant_request, render_one(my_grant_request, GrantRequestView, "grant_request.json"))
+  defp maybe_put_my_grant_request(json, %{my_grant_request: [%{} = my_grant_request]}),
+    do:
+      Map.put_new(
+        json,
+        :my_grant_request,
+        render_one(my_grant_request, GrantRequestView, "grant_request.json")
+      )
+
   defp maybe_put_my_grant_request(json, _), do: json
 
-  defp maybe_put_my_grants(json, %{my_grants: [_ | _] = my_grants}), do:
-    Map.put_new(json, :my_grants, render_many(my_grants, GrantView, "grant.json"))
+  defp maybe_put_my_grants(json, %{my_grants: [_ | _] = my_grants}),
+    do: Map.put_new(json, :my_grants, render_many(my_grants, GrantView, "grant.json"))
+
   defp maybe_put_my_grants(json, _), do: json
 
-  defp maybe_put_grant_user_permissions(json, %{user_permissions: %{} = user_permissions}), do:
-    Map.put_new(json, :user_permissions, user_permissions)
+  defp maybe_put_grant_user_permissions(json, %{user_permissions: %{} = user_permissions}),
+    do: Map.put_new(json, :user_permissions, user_permissions)
+
   defp maybe_put_grant_user_permissions(json, _), do: json
 
   defp add_children(data_structure_json, data_structure),