+++ test/td_dd/grants_test.exs
@@ -44,13 +44,17 @@ defmodule TdDd.GrantsTest do
 
   describe "list_active_grants!/1" do
     test "returns the grants of a list of data_structures by user", %{user_id: user_id} do
-      %{id: id, data_structure_id: data_structure_id} = insert(:grant, user_id: user_id, end_date: nil)
+      %{id: id, data_structure_id: data_structure_id} =
+        insert(:grant, user_id: user_id, end_date: nil)
+
       insert(:grant, data_structure_id: data_structure_id)
       insert(:grant, user_id: user_id)
-      assert [%{id: ^id}] = Grants.list_active_grants(%{
-        data_structure_ids: [data_structure_id],
-        user_id: user_id,
-      })
+
+      assert [%{id: ^id}] =
+               Grants.list_active_grants(%{
+                 data_structure_ids: [data_structure_id],
+                 user_id: user_id
+               })
     end
   end