+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 - [TD-5732] Fixed reindex all implementation versions when add/delete links
 - [TD-5442] Fix hard delete implementation versions
 
+### Fixed
+
+- [TD-5680] Fix Children Data Structure Versions populate their classes correctly
+
 ## [5.5.1] 2023-04-18
 
 ### Fixed